// server-backup/front-end/src/service/api/student/type.ts
export type StudentInterface = {
  id: number;
  userId: number;
  studentId: string;
  name: string;
  middleName: string | null;
  surname: string;
  gpax: number;
  phoneNumber: string;
  email: string;
  picture: File | string | null;
  curriculumId: string;
  programId: string;
  facultyId: string;
  majorId: string;
  campusId: string;
  faculty: {
    id: number;
    campusId: number;
    facultyNameEn: string;
    facultyNameTh: string;
  };
  program: {
    id: number;
    facultyId: number;
    programNameEn: string;
    programNameTh: string;
    createdAt: string;
    updatedAt: string;
  };
};
export type StudentDTO = {
  user_id: number;
  student_id: string;
  name: string;
  middle_name: string | null;
  surname: string;
  gpax: number;
  phone_number: string;
  email: string;
  picture: File | string | null;
  major_id: number;
  program_id: number;
  curriculum_id: number;
  faculty_id: number;
  campus_id: number;
};